pyproject.toml
[project]
name = "openwhisper-desktop"
version = "0.1.0"
description = "A messaging application with a focus on privacy and security. Written in Python and Django."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "channels[daphne]>=4.3.2",
    "channels-redis>=4.3.0",
    "django==5.2.13",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=12.2.0",
    "psycopg[binary]>=3.3.3",
    "redis>=7.4.0",
    "django-environ>=0.13.0",
    "django-filter>=25.2",
    "django-tailwind>=4.4.2",
]

[dependency-groups]
dev = [
    "channels[types]>=4.3.2",
    "django-debug-toolbar>=6.3.0",
    "django-stubs>=6.0.3",
    "django-tailwind[cookiecutter,honcho,reload]>=4.4.2",
    "djlint>=1.36.4",
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "pytest-django>=4.12.0",
    "ruff>=0.15.12",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "openwhisper.settings.tests"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

[tool.mypy]
python_version = "3.13"
plugins = ["mypy_django_plugin.main"]
disable_error_code = ["import-untyped"]

[tool.django-stubs]
django_settings_module = "openwhisper.settings.local"
